+++ rana_qgis_plugin/utils/api.py
@@ -6,7 +6,7 @@ from rana_qgis_plugin.auth import get_authcfg_id
 from rana_qgis_plugin.communication import UICommunication
 from rana_qgis_plugin.constant import COGNITO_USER_INFO_ENDPOINT
 from rana_qgis_plugin.network_manager import NetworkManager
-from rana_qgis_plugin.utlis.settings import api_url, get_tenant_id
+from rana_qgis_plugin.utils.settings import api_url, get_tenant_id
 
 
 class UserInfo(TypedDict):
@@ -731,7 +731,7 @@ def upload_publication_version(publication_id: str, publication_version: dict):
             detail = network_manager.content.get("detail", {})
             created_by = detail.get("created_by_name", "unknown")
             if "created_at" in detail:
-                from rana_qgis_plugin.utlis.time import (
+                from rana_qgis_plugin.utils.time import (
                     convert_timestamp_str_to_relative_time,
                 )
