+++ rana_qgis_plugin/utils/lizard.py
