+++ rana_qgis_plugin/utils/settings.py
