+++ rana_qgis_plugin/utils/spatial.py
