+++ rana_qgis_plugin/widgets/file_view.py
@@ -28,19 +28,19 @@ from qgis.PyQt.QtWidgets import (
 from rana_qgis_plugin.auth_3di import has_3di_authcfg
 from rana_qgis_plugin.constant import SUPPORTED_DATA_TYPES
 from rana_qgis_plugin.simulation.threedi_calls import ThreediCalls
-from rana_qgis_plugin.utlis.api import (
+from rana_qgis_plugin.utils.api import (
     get_tenant_file_descriptor,
     get_tenant_project_file,
     get_threedi_schematisation,
 )
-from rana_qgis_plugin.utlis.generic import (
+from rana_qgis_plugin.utils.generic import (
     NumericItem,
     display_bytes,
     get_file_icon_name,
     get_threedi_api,
 )
-from rana_qgis_plugin.utlis.spatial import get_bbox_area_in_m2
-from rana_qgis_plugin.utlis.time import (
+from rana_qgis_plugin.utils.spatial import get_bbox_area_in_m2
+from rana_qgis_plugin.utils.time import (
     format_activity_timestamp,
     format_activity_timestamp_str,
     parse_timestamp_str,
