+++ rana_qgis_plugin/widgets/settings_dialog.py
@@ -14,9 +14,9 @@ from qgis.PyQt.QtWidgets import (
 )
 
 from rana_qgis_plugin.constant import PLUGIN_NAME
-from rana_qgis_plugin.utlis.api import get_frontend_settings
-from rana_qgis_plugin.utlis.generic import is_writable
-from rana_qgis_plugin.utlis.settings import (
+from rana_qgis_plugin.utils.api import get_frontend_settings
+from rana_qgis_plugin.utils.generic import is_writable
+from rana_qgis_plugin.utils.settings import (
     base_url,
     hcc_working_dir,
     rana_cache_dir,
