+++ rana_qgis_plugin/workers/upload.py
@@ -9,7 +9,7 @@ from qgis.PyQt.QtCore import (
     pyqtSlot,
 )
 
-from rana_qgis_plugin.utlis.api import (
+from rana_qgis_plugin.utils.api import (
     finish_file_upload,
     get_tenant_project_file,
     start_file_upload,
