+++ tests/test_utils_spatial.py
@@ -2,7 +2,7 @@ import math
 
 import pytest
 
-from rana_qgis_plugin.utlis.spatial import get_bbox_area_in_m2
+from rana_qgis_plugin.utils.spatial import get_bbox_area_in_m2
 
 
 @pytest.mark.parametrize(
